ENH: update route file-generator structure
